backend/requirements.txt
# ============================================
# Core Dependencies
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# Database Drivers
# ============================================
neo4j==5.16.0
psycopg[pool]==3.1.18
elasticsearch==8.12.0
weaviate-client==3.26.0
redis==5.0.1
minio==7.2.3

# ============================================
# LLM Providers
# ============================================
anthropic==0.17.0
openai==1.12.0
langchain==0.1.6
langchain-anthropic==0.1.1
langchain-openai==0.0.5
langgraph==0.0.20
# ============================================
# NLP & ML
# ============================================
spacy==3.7.2
transformers==4.37.2
torch==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# ============================================
# Web Scraping & Collection
# ============================================
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.26.0
tweepy==4.14.0
praw==7.7.1

# Tor support
PySocks==1.7.1
stem==1.8.2

# ============================================
# Utilities
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1
tenacity==8.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0  # For testing HTTP requests

# ============================================
# Development
# ============================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
